# pkg/gateway/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lenny Gateway API",
    "description": "Lenny session, blob, and admin endpoints. See spec/15_external-api-surface.md.",
    "version": "0.1.0",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "/", "description": "Gateway root" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "§10.2 OIDC/OAuth 2.1 access token."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "§15.1 line 972 / §15.2.1 line 1376 REST↔MCP error envelope. Every gateway error response carries this shape verbatim; SDK generators must surface category and retryable so callers can distinguish a transient network blip from a permanent validation failure.",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "category", "message", "retryable"],
            "properties": {
              "code":      { "type": "string", "description": "§15.1 canonical error code (e.g. VALIDATION_ERROR, UNAUTHORIZED, MCP_PROTOCOL_VERSION_RETIRED)." },
              "category":  { "type": "string", "enum": ["VALIDATION", "AUTH", "AUTHZ", "NOT_FOUND", "CONFLICT", "TRANSIENT", "RATE_LIMIT", "INTERNAL", "UNAVAILABLE", "TIMEOUT"], "description": "§15.1 line 972 — coarse classification used by SDK retry logic." },
              "message":   { "type": "string", "description": "Human-readable failure summary." },
              "retryable": { "type": "boolean", "description": "§15.1 line 972 — true when the client should retry, with backoff bounded by suggestedRetryAfter when present." },
              "details":   { "type": "object", "additionalProperties": true, "description": "Optional structured per-error context (fields, runbook, suggestedRetryAfter, …)." }
            }
          }
        }
      },
...
      "Session": {
        "type": "object",
        "required": ["id", "tenantId", "state", "createdAt", "updatedAt"],
        "properties": {
          "id":         { "type": "string" },
          "tenantId":   { "type": "string" },
          "userId":     { "type": "string" },
          "runtimeRef": { "type": "string" },
          "environment": { "type": "string" },
          "state": {
            "type": "string",
            "description": "§15.1 session state. `resuming` is an internal transient that the GET /v1/sessions/{id} envelope normalises to `resume_pending` → `running`; it appears only on the §10.4 coordinator-handoff `status_change` SSE frame and is included in the enum for SSE wire compatibility.",
            "enum": ["created","finalizing","ready","starting","running","input_required","suspended","resume_pending","resuming","awaiting_client_action","completed","failed","cancelled","expired"]
          },
          "createdAt":  { "type": "string", "format": "date-time" },
          "updatedAt":  { "type": "string", "format": "date-time" },
          "failureClass": {
            "type": "string",
            "enum": ["runtime_failure","starting_timeout","budget_keys_expired","workspace_seal_timeout","derive_failure"]
          },
          "metadata": {
            "type": "object",
            "description": "§7.1 line 6 client-supplied CreateSession metadata, a flat string→string map echoed verbatim across the session lifetime.",
            "additionalProperties": { "type": "string" }
          },
          "retryPolicy": { "$ref": "#/components/schemas/RetryPolicy" }
        }
      },
      "CreateSessionRequest": {
        "type": "object",
        "required": ["runtimeRef"],
        "properties": {
          "runtimeRef":       { "type": "string" },
          "userId":           { "type": "string" },
          "environment":      { "type": "string" },
          "workspacePlan":    { "type": "object", "additionalProperties": true },
          "isolationProfile": { "type": "string", "enum": ["standard","sandboxed","microvm"] },
          "metadata": {
            "type": "object",
            "description": "§7.1 line 6 client annotations preserved verbatim for the session lifetime. Must be a flat string→string map; non-string values reject with 400 VALIDATION_ERROR.",
            "additionalProperties": { "type": "string" }
          },
          "retryPolicy": { "$ref": "#/components/schemas/RetryPolicy" }
        }
      },
      "RetryPolicy": {
        "type": "object",
        "description": "§7.3 client-supplied retry policy. The gateway clamps each populated field against the deployer caps at admission; a missing/zero value falls through to the cap as the effective value. Negative values reject with 400 VALIDATION_ERROR.",
        "properties": {
          "mode": { "type": "string", "enum": ["auto_then_client", "client_only"], "description": "§7.3 line 382 retry mode. Defaults to auto_then_client." },
          "maxRetries": { "type": "integer", "minimum": 0, "description": "§7.3 line 383 auto-retry budget." },
          "retryableFailures": { "type": "array", "items": { "type": "string" }, "description": "§7.3 line 384 retryable failure class advisory list." },
          "nonRetryableFailures": { "type": "array", "items": { "type": "string" }, "description": "§7.3 line 385 non-retryable failure class advisory list." },
          "maxSessionAgeSeconds": { "type": "integer", "minimum": 0, "description": "§7.3 line 389 per-session total-lifetime cap, clamped against the deployer cap." },
          "maxResumeWindowSeconds": { "type": "integer", "minimum": 0, "description": "§7.3 line 390 resume-window timer, clamped against the deployer cap." }
        }
      },
      "CreateSessionResponse": {
        "type": "object",
        "required": ["id", "tenantId", "state", "uploadToken", "sessionIsolationLevel"],
        "allOf": [
          { "$ref": "#/components/schemas/Session" },
          {
            "type": "object",
            "properties": {
              "uploadToken": { "type": "string" },
              "sessionIsolationLevel": {
                "type": "object",
                "required": ["executionMode", "isolationProfile", "podReuse", "residualStateWarning", "conversationContinuity"],
                "properties": {
                  "executionMode":          { "type": "string", "enum": ["session","service"] },
                  "isolationProfile":       { "type": "string", "enum": ["standard","sandboxed","microvm"] },
                  "podReuse":               { "type": "boolean" },
                  "scrubPolicy":            { "type": "string" },
                  "residualStateWarning":   { "type": "boolean" },
                  "conversationContinuity": { "type": "string", "enum": ["platform","none"], "description": "§7.1 — `platform` for session mode (the platform preserves conversation context across messages for the session's lifetime); `none` for service mode (each message routes to any ready replica and clients re-inject context)." }
                }
              }
            }
          }
        ]
      },
      "DeriveRequest": {
        "type": "object",
        "properties": {
          "runtimeRef":              { "type": "string" },
          "targetPool":              { "type": "string" },
          "targetIsolationProfile":  { "type": "string", "enum": ["standard","sandboxed","microvm"] },
          "userId":                  { "type": "string" },
          "allowStale":              { "type": "boolean", "default": false },
          "allowIsolationDowngrade": { "type": "boolean", "default": false },
          "ticketId":                { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/sessions": {
      "post": {
        "operationId": "postV1Sessions",
        "summary": "Create a session",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSessionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateSessionResponse" }
              }
            }
          },
          "400": { "description": "Validation failure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "operationId": "getV1Sessions",
        "summary": "List sessions",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "state",      "in": "query", "schema": { "type": "string" } },
          { "name": "runtime",    "in": "query", "schema": { "type": "string" } },
          { "name": "failureClass", "in": "query", "schema": { "type": "string" } },
          { "name": "cursor", "in": "query", "description": "Opaque cursor from a previous page's response. Omit for the first page.", "schema": { "type": "string" } },
          { "name": "limit",  "in": "query", "description": "Items per page. Minimum 1, maximum 200; values outside the range are clamped. Default 50.", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } },
          { "name": "sort",   "in": "query", "description": "Sort field and direction as field:asc|desc. Supported fields: created_at, updated_at. Default created_at:desc.", "schema": { "type": "string", "default": "created_at:desc" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items", "hasMore"],
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Session" } },
                    "cursor": { "type": "string", "nullable": true, "description": "§15.1 line 1251 — opaque cursor for the next page; null when no more results." },
                    "hasMore": { "type": "boolean", "description": "§15.1 line 1251 — true when additional pages exist beyond this one." },
                    "total": { "type": "integer", "description": "§15.1 line 1252 — total match count across all pages; present only when cheaply computable." }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/environments/{name}/sessions": {
      "post": {
        "operationId": "postV1EnvironmentsByNameSessions",
        "summary": "Create a session in an environment",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSessionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateSessionResponse" }
              }
            }
          },
          "400": { "description": "Validation failure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/v1/sessions/{id}": {
      "get": {
        "operationId": "getV1SessionsById",
        "summary": "Get session",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } } },
          "404": { "description": "Not found",  "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "operationId": "deleteV1SessionsById",
        "summary": "Cancel session",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Cancelled", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } } }
        }
      }
    },
...
    "/v1/sessions/{id}/derive": {
      "post": {
        "operationId": "postV1SessionsByIdDerive",
        "summary": "Derive a new session from this session's workspace snapshot",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeriveRequest" } } }
        },
        "responses": {
          "201": { "description": "Created" },
          "409": { "description": "DERIVE_ON_LIVE_SESSION" },
          "422": { "description": "ISOLATION_MONOTONICITY_VIOLATED" }
        }
      }
    },
...
    "/v1/admin/sessions/{id}": {
      "get": {
        "operationId": "getV1AdminSessionsById",
        "summary": "Read-through a session by its global id",
        "x-lenny-mcp-tool": "admin.get_session",
        "x-lenny-scope": "admin.sessions.read",
        "x-lenny-required-role": "platform-admin",
        "x-lenny-category": "security",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}